chore: update contributing guide to mention extensions
The contributing guide has an entry on extensions - but it is out of date, and buried
enough that I didn't find it on my first pass.

This creates a section to add some visibility (too high up? It felt like it should be
before contributing to Podman Desktop itself), along with a bit of text to mention
what you can do with extensions.

Part of #10715.

Signed-off-by: Tim deBoer <git@tdeboer.ca>