feat: kube pods page
Adds a Kubernetes Pods page, similar to Deployments/Services/etc, but with the
same columns and actions as the Kubernetes pods in the current Pods page.

Given the number of files just to add the page, this commit intentionally does not
include a details page. That will come in following PRs.

Since we already have a PodsList in src/lib/pod, I've put this in /src/lib/kube/pods,
not sure if there is a better location or we should start organizing Kube folders
differently.

Main page and columns updated to Svelte 5, but not PodActions. Updating this page
would force moving to Svelte 5 event callbacks, which will affect most tables and
should be done as a separate task.

Part of #8818.

Signed-off-by: Tim deBoer <git@tdeboer.ca>