chore: kubernetes navigation store mocking
When we added Pods to the Kubernetes navigation, we used the newer vi.mocking.
This just updates the existing tests to use the same mocking.

Note: the actual objects being mocked are irrelevant for the tests, they would
still pass if they used [{},{}]. Instead of switching to that I've cleaned
up the objects and array name in case we ever change
kubernetesRegisterGetCurrentContextResources such that the tests rely on the
correct type of resource.

Part of #10903.

Signed-off-by: Tim deBoer <git@tdeboer.ca>