chore: common kubernetes ui objects
Creates a new KubernetesObjectUI and uses this as the parent for all Kubernetes
UI objects like DeploymentUI, etc., with an optional kind property (*). This improves
typing and in future PRs will allow us to have a shared Kubernetes Name column,
map from kind to plural (url), and remove Kind.isResource() from
KubernetesObjectsList.

(*) Although we could make more properties required over time only name + status
are required for now to keep the change manageable. E.g. enforcing kind would
require changes in 35 additional test files. Look at this as the first step; we
can enforce columns or evolve from here.

UI objects are using hard-coded kinds like `kind: 'Pod'` for now, because the kind
is not set on our resources. As we move to the experimental mode this shouldn't be
required and we could use `kind: pod.kind ?? ''`. Likewise, the kind doesn't
strictly need to be set on the test objects yet but I did this so they'll be correct
later.

Signed-off-by: Tim deBoer <git@tdeboer.ca>