chore: remove kubernetes pod watcher
After #11323, the Kubernetes client listPods() function is no longer being called.
This function can be removed, along with the Kube watcher that existed solely to
provide the data for it, and the pod-event that is no longer used through this
context (it is fired from other cases - just not this one).

Another part of #8819.

Signed-off-by: Tim deBoer <git@tdeboer.ca>