chore: kubernetes namespaced object ui
Split the interface to have a KubernetesNamespacedObjectUI where namespace is
required. The new interface is used by all children except nodes.

Node happens to be the only child that didn't have a selected property, so I
also moved it to namespaced object and made it required so that we could
remove both namespace and selected from all children.

Signed-off-by: Tim deBoer <git@tdeboer.ca>