chore: port forwarding should open kubernetes pod page
The port forwarding feature was relying on the list of pods from the Pods store,
(which used the Kubernetes watcher), then opened the regular Pods page.

The watcher could be out of sync/delayed; we should be using the Kubernetes
informer and opening the new Kubernetes > Pods page instead.

Part of #8819.

Signed-off-by: Tim deBoer <git@tdeboer.ca>